pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "doc-sanitizer"
version = "0.2.0"
description = "Local-first document masking and rehydration with an HTTPS API and library mode."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "cogniflow-ai" }
]
keywords = ["masking", "redaction", "obfuscation", "PII", "docx", "pptx", "xlsx"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]

dependencies = [
    "flask>=3.0",
    "python-docx>=1.1",
    "python-pptx>=0.6",
    "openpyxl>=3.1",
    "cryptography>=42",
    "keyring>=24",
]

[project.optional-dependencies]
build = ["pyinstaller>=6.6"]
test  = ["pytest>=8"]

[project.scripts]
doc-sanitizer = "doc_sanitizer.cli:main"

[project.urls]
Homepage   = "https://github.com/cogniflow-ai/doc-sanitizer"
Repository = "https://github.com/cogniflow-ai/doc-sanitizer"
Issues     = "https://github.com/cogniflow-ai/doc-sanitizer/issues"

[tool.setuptools]
packages = ["doc_sanitizer"]
include-package-data = true

[tool.setuptools.package-data]
doc_sanitizer = ["templates/*.html", "templates/partials/*.html", "static/*"]
